+++ crates/cxx-qt-lib/src/core/mod.rs
@@ -97,6 +97,9 @@ pub use qpointf::QPointF;
 mod qurl;
 pub use qurl::QUrl;
 
+mod quuid;
+pub use quuid::{QUuid, QUuidVariant, QUuidVersion};
+
 mod qvariant;
 pub use qvariant::{QVariant, QVariantValue};